shared start fn
