+++ packages/provider/src/index.ts
@@ -20,6 +20,7 @@ export * from "./api/verify.js";
 export * from "./api/ja4Middleware.js";
 export * from "./api/public.js";
 export * from "./api/domainMiddleware.js";
+export * from "./api/startProviderApi.js";
 export * from "./schedulers/captchaScheduler.js";
 export * from "./schedulers/getClientList.js";
 export * from "./schedulers/setClientEntropy.js";
